feat: add GM simulation core and tests; refactor gm_sim to core/gm/sim/gm_sim_core.js, expose runGmSim, add test script, CI test, and new GM Roadmap + workflow tweaks
